refactor export function to utilize hook
