+++ src/PrintOrder/PrintOrder.test.js
@@ -3,13 +3,8 @@ import { render, screen } from '@folio/jest-config-stripes/testing-library/react
 import { order, orderLine } from 'fixtures';
 import { PrintOrder } from './PrintOrder';
 
-jest.mock('@folio/stripes/core', () => ({
-  ...jest.requireActual('@folio/stripes/core'),
-  useOkapiKy: jest.fn(() => ({})),
-  useStripes: jest.fn(() => ({
-    currency: 'USD',
-  })),
-  stripesConnect: jest.fn(component => component),
+jest.mock('./hooks', () => ({
+  usePOLineTotalEstimatedPrice: jest.fn().mockReturnValue({ getPOLineTotalEstimatedPrice: jest.fn() }),
 }));
 jest.mock('./ComponentToPrint', () => jest.fn().mockReturnValue('ComponentToPrint'));
 jest.mock('../common/ExportSettingsModal/utils/getExportData', () => ({
