+++ src/PrintOrder/utils.test.js
@@ -1,17 +1,7 @@
 import { order } from '../../test/jest/fixtures';
 import {
   getOrderPrintData,
-  getCurrencyRate,
-  getPOLineTotalEstimatedPrice,
 } from './utils';
-
-const poLine = {
-  currency: 'EUR',
-  exchangeRate: 1,
-  poLineEstimatedPrice: 1,
-  quantityElectronic: 0,
-  quantityPhysical: 1,
-};
 
 describe('utils', () => {
   it('should call `getOrderPrintData`', () => {
@@ -24,85 +14,5 @@ describe('utils', () => {
     getOrderPrintData(ky, order);
 
     expect(ky.get).toHaveBeenCalled();
-  });
-
-  it('should call `getCurrencyRate`', async () => {
-    const ky = {
-      get: jest.fn(() => ({
-        json: () => Promise.resolve({
-          exchangeRate: 1,
-        }),
-      })),
-    };
-
-    const resp = await getCurrencyRate(ky, order);
-
-    expect(resp).toEqual(1);
-  });
-
-  describe('getPOLineTotalEstimatedPrice', () => {
-    let ky;
-
-    beforeEach(() => {
-      jest.clearAllMocks();
-
-      ky = {
-        get: jest.fn(() => ({
-          json: () => Promise.resolve({
-            exchangeRate: 2,
-          }),
-        })),
-      };
-    });
-
-    it('should call `getPOLineTotalEstimatedPrice`', async () => {
-      const resp = await getPOLineTotalEstimatedPrice({ ky, poLine, systemCurrency: 'USD' });
-
-      expect(resp).toEqual({
-        totalItems: 1,
-        totalEstimatedPrice: 1,
-      });
-      expect(ky.get).not.toHaveBeenCalled();
-    });
-
-    it('should call `getPOLineTotalEstimatedPrice` with `getCurrencyRate`', async () => {
-      const resp = await getPOLineTotalEstimatedPrice({
-        ky,
-        poLine: {
-          ...poLine,
-          exchangeRate: null,
-        },
-        systemCurrency: 'USD',
-      });
-
-      expect(resp).toEqual({
-        totalItems: 1,
-        totalEstimatedPrice: 2,
-      });
-
-      expect(ky.get).toHaveBeenCalled();
-    });
-
-    it('should call `getPOLineTotalEstimatedPrice` with `getCurrencyRate` with reject', async () => {
-      const resp = await getPOLineTotalEstimatedPrice({
-        ky: {
-          get: jest.fn(() => ({
-            json: () => Promise.reject(),
-          })),
-        },
-        poLine: {
-          ...poLine,
-          exchangeRate: null,
-        },
-        systemCurrency: 'USD',
-      });
-
-      expect(resp).toEqual({
-        totalItems: 1,
-        totalEstimatedPrice: 1,
-      });
-
-      expect(ky.get).not.toHaveBeenCalled();
-    });
   });
 });
